`revokeToken()` returns void, schedule testing
