+++ .github/workflows/test.yml
@@ -6,6 +6,8 @@ on:
   pull_request:
     branches:
       - "**"
+  schedule:
+    - cron: "45 10 12 * *"
 
 jobs:
   build:
