+++ .npmignore
@@ -5,6 +5,6 @@
 lib/
 docs/
 tests/
-yarn.lock
+package-lock.json
 tsconfig.json
 CONTRIBUTING.md
